// web/src/api/generated/models/PersonRelationship.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PersonRelationship = {
    CHILD: 'CHILD',
    PARENT: 'PARENT',
    SIBLING: 'SIBLING',
    FRIEND: 'FRIEND',
    OTHER: 'OTHER'
} as const;
export type PersonRelationship = typeof PersonRelationship[keyof typeof PersonRelationship];


export function instanceOfPersonRelationship(value: any): boolean {
    for (const key in PersonRelationship) {
        if (Object.prototype.hasOwnProperty.call(PersonRelationship, key)) {
            if (PersonRelationship[key as keyof typeof PersonRelationship] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PersonRelationshipFromJSON(json: any): PersonRelationship {
    return PersonRelationshipFromJSONTyped(json, false);
}

export function PersonRelationshipFromJSONTyped(json: any, ignoreDiscriminator: boolean): PersonRelationship {
    return json as PersonRelationship;
}

export function PersonRelationshipToJSON(value?: PersonRelationship | null): any {
    return value as any;
}

export function PersonRelationshipToJSONTyped(value: any, ignoreDiscriminator: boolean): PersonRelationship {
    return value as PersonRelationship;
}